backend/requirements.txt
# Backend dependencies
django>=5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
djangorestframework-simplejwt[crypto]>=5.3.0
drf-nested-routers>=0.93.3
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0